is over...

+++ app/src/main/java/com/sarpjfhd/cashwise/fragments/TransactionListFragment.kt
@@ -8,6 +8,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -16,6 +17,7 @@ import com.sarpjfhd.cashwise.MainViewModel
 import com.sarpjfhd.cashwise.R
 import com.sarpjfhd.cashwise.models.*
 import com.sarpjfhd.cashwise.navigateSafe
+import okhttp3.ResponseBody
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -81,16 +83,18 @@ class TransactionListFragment(private var expenses: MutableList<Expense>, privat
     }
 
     fun deleteTransaction(apiServiceInterface: ServiceCallback, transactionId: Int) {
+        val tdata = TransactionData()
+        tdata.idBD = transactionId
         val service: Service = RestEngine.getRestEngine().create(Service::class.java)
-        val result: Call<Boolean> = service.deleteTransaction(transactionId)
+        val result: Call<ResponseBody> = service.deleteTransaction(tdata)
 
-        result.enqueue(object: Callback<Boolean> {
-            override fun onFailure(call: Call<Boolean>, t: Throwable) {
+        result.enqueue(object: Callback<ResponseBody> {
+            override fun onFailure(call: Call<ResponseBody>, t: Throwable) {
                 Toast.makeText(requireContext(), "Error obteniendo los mensajes", Toast.LENGTH_LONG).show()
             }
 
-            override fun onResponse(call: Call<Boolean>, response: Response<Boolean>) {
-                apiServiceInterface.onSuccess(response.body()!!)
+            override fun onResponse(call: Call<ResponseBody>, response: Response<ResponseBody>) {
+                apiServiceInterface.onSuccess(true)
             }
         })
     }
@@ -104,11 +108,27 @@ class TransactionListFragment(private var expenses: MutableList<Expense>, privat
         findNavController().navigateSafe(action)
     }
 
-    override fun onDeleteClick(transactionId: Int) {
-        deleteTransaction(object: ServiceCallback {
-            override fun onSuccess(result: Boolean) {
-                TODO("Not yet implemented")
+    override fun onDeleteClick(transactionId: Int, transactions: MutableList<Transaction>, index: Int) {
+        val options =  arrayOf("First Item", "Second Item", "Third Item")
+        val singleChoiceDialog = AlertDialog.Builder(requireContext())
+            .setTitle("¿Esta seguro que quiere eliminar esta transacción?")
+            .setSingleChoiceItems(options, 0){dialogInterface, i ->
+
+                Toast.makeText(context, "You clicked on ${options[i]}", Toast.LENGTH_SHORT).show()
+
             }
-        }, transactionId)
+            .setPositiveButton("Aceptar"){ _,_ ->
+                deleteTransaction(object: ServiceCallback {
+                    override fun onSuccess(result: Boolean) {
+                        transactionAdapter?.deleteItem(index)
+                    }
+                }, transactionId)
+            }
+            .setNegativeButton("Cancelar"){ _,_ ->
+
+                Toast.makeText(context,"Cancel",Toast.LENGTH_SHORT).show()
+            }.create()
+
+        singleChoiceDialog.show()
     }
 }
